+++ src/control_plane/engine/mod.rs
@@ -1,3 +1,4 @@
+mod bollard_engine_adapter;
 mod container_create_options;
 mod container_id;
 mod container_lifecycle;
@@ -8,7 +9,7 @@ mod resource_kind;
 
 pub(crate) use container_create_options::ContainerCreateOptions;
 pub(crate) use container_id::ContainerId;
-pub(crate) use container_lifecycle::ContainerLifecycle;
+pub(crate) use container_lifecycle::{ContainerLifecycle, EngineFuture};
 pub(crate) use container_state::ContainerState;
 pub(crate) use engine_error::EngineError;
 pub(crate) use managed_resource_metadata::ManagedResourceMetadata;
@@ -16,3 +17,4 @@ pub(crate) use resource_kind::ResourceKind;
 
 #[cfg(test)]
 mod tests;
+pub(crate) use bollard_engine_adapter::BollardEngineAdapter;